Backward NaN-lse before ignore_index + drop wrapper pre-cast for PR #682
A wide invalid label that narrows to ignore_index after astype(mx.int32)
(e.g. 2**32 - 100 -> -100 with ignore_index=-100) used to produce a NaN
forward loss but a zero backward gradient. Move the NaN-lse guard before
the ignore_index check in both the Metal dlogits kernel and the Python
fallback so the gradient also poisons. Refuse the implicit zeroing path
on rows whose lse is NaN.

Also drop the .astype(mx.int32) pre-cast from the four CCE wrapper call
sites (text dense / quantized; VLM dense / quantized). Casting before the
runtime narrowed wide labels to valid class ids or ignore_index, slipping
past the new original-dtype validation in _forward_chunked_fused_finalize.
The runtime now performs that narrow internally only after the validity
masks are built from the original dtype.

New regression: test_runtime_cce_int64_wrap_to_ignore_index_poisons_gradients.

+++ unsloth_zoo/mlx/utils.py
@@ -361,7 +361,7 @@ def make_cce_loss_fn(model):
             masked_targets = mx.where(mask, targets, -100)
             ntoks = mask.sum()
             hidden_flat = hidden.reshape((-1, hidden.shape[-1]))
-            targets_flat = masked_targets.reshape((-1,)).astype(mx.int32)
+            targets_flat = masked_targets.reshape((-1,))  # runtime CCE validates dtype before narrowing
             loss = rt_cce(hidden_flat, w, sc, bi, targets_flat)
             loss = loss.astype(mx.float32).sum() / _safe_token_denominator(ntoks)
             return loss, ntoks
@@ -394,7 +394,7 @@ def make_cce_loss_fn(model):
             masked_targets = mx.where(mask, targets, -100)
             ntoks = mask.sum()
             hidden_flat = hidden.reshape((-1, hidden.shape[-1]))
-            targets_flat = masked_targets.reshape((-1,)).astype(mx.int32)
+            targets_flat = masked_targets.reshape((-1,))  # runtime CCE validates dtype before narrowing
             loss = rt_cce(hidden_flat, w, targets_flat)
             loss = loss.astype(mx.float32).sum() / _safe_token_denominator(ntoks)
             return loss, ntoks
@@ -1542,7 +1542,7 @@ def make_vlm_cce_loss_fn(model, assistant_token_id=0):
             # gradients (see runtime_cce.py VJP), so stop_gradient is
             # redundant here even when the LM head is frozen.
             hidden_flat = hidden.reshape((-1, hidden.shape[-1]))
-            targets_flat = masked_targets.reshape((-1,)).astype(mx.int32)
+            targets_flat = masked_targets.reshape((-1,))  # runtime CCE validates dtype before narrowing
             loss = rt_cce(hidden_flat, w, sc, bi, targets_flat)
             loss = loss.astype(mx.float32).sum() / _safe_token_denominator(ntoks)
             return loss, ntoks
@@ -1563,7 +1563,7 @@ def make_vlm_cce_loss_fn(model, assistant_token_id=0):
             if _skip_weight_grad:
                 w = mx.stop_gradient(w)
             hidden_flat = hidden.reshape((-1, hidden.shape[-1]))
-            targets_flat = masked_targets.reshape((-1,)).astype(mx.int32)
+            targets_flat = masked_targets.reshape((-1,))  # runtime CCE validates dtype before narrowing
             loss = rt_cce(hidden_flat, w, targets_flat)
             loss = loss.astype(mx.float32).sum() / _safe_token_denominator(ntoks)
             return loss, ntoks
